ImageMagick updated to the latest version, ImageMagickHelper fixed/updated
Former-commit-id: 3c56d8784f7e19a69c637a6be1ae48731fa84aea

+++ Sources/Internal/Render/RHI/GLES2/rhi_CommandBufferGLES2.cpp
@@ -472,7 +472,7 @@ gles2_CommandBuffer_SetMarker(Handle cmdBuf, const char* text)
         cb->text->Initialize(64 * 1024);
     }
 
-    int len = strlen(text);
+    size_t len = strlen(text);
     char* txt = (char*)cb->text->Alloc(len / sizeof(float) + 2);
 
     memcpy(txt, text, len);
